refactored user registration to auth service
